+++ routes/index.js
@@ -1,13 +1,13 @@
 const router = require('express').Router();
-const apiRoutes = require('./api');
-const htmlRoutes = require('./html-routes')
-const spotifyRoutes = require('./spotify-routes')
 
-router.use('', spotifyRoutes)
+const apiRoutes = require('./api');
+const mainRoutes = require('./mainRoutes');
+
 router.use('/api', apiRoutes);
-router.use('', htmlRoutes)
+router.use('/', mainRoutes);
+
 router.use((req, res) => {
-    res.send("<h1>Wrong Route!</h1>")
+    res.status(404).end();
 });
 
 module.exports = router;
