+++ server.js
@@ -4,11 +4,11 @@ const path = require('path');
 const exphbs = require('express-handlebars');
 const sequelize = require('./config/connections')
 const session = require('express-session');
-const db = require('./models');
-const port = process.env.PORT || 8888;
+const routes = require('./routes');
 // const SequilzeStore = require('connect-session-sequelize')(session.Store);
 
 const app = express()
+const port = process.env.PORT || 8888;
 
 //cookie session requirements i think NEEDS TO BE FINISHED SESSION IS A BIG THING WE NEED DONE
 // const sess = {
@@ -32,21 +32,16 @@ app.use(
 app.use(express.json())
 app.use(express.urlencoded({ extended: true }))
 
-// .env file access
-require('dotenv').config()
-
 // setting Handlebars as the default template engine.
 app.engine('handlebars', exphbs({ defaultLayout: 'main' }));
 app.set('view engine', 'handlebars');
 
-// path setting
-const publicPath = path.resolve(__dirname, "public");
-
-app.use(express.static(publicPath));
+// STATIC FILE PATH
+// const publicPath = path.resolve(__dirname, "public");
+app.use(express.static('public'));
 
 // ROUTES
-// ===========================================================
-app.use(require('./routes'));
+app.use(routes);
 
 // SERVER LISTEN
 sequelize.sync({ force: false }).then(() => {
