+++ views/landing.handlebars
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 <nav class="navbar">
   <h1>Offline Beats</h1>
   <h4>
@@ -10,52 +9,42 @@
     <!-- Modal content -->
     <div class="modal-content">
 
-      
-
       <form action="/">
-      <h2>Login Page</h2>
+        <h2>Login Page</h2>
         <label for="Uname">Username:</label><br />
         <input type="text" id="Uname" name="Uname" value="Username" /><br />
         <label for="password">Password:</label><br />
-        <input type="text" id="password" name="Password" value="password" /><br
-        /><br />
+        <input type="text" id="password" name="Password" value="password" /><br /><br />
         <input id="subbmit" type="submit" value="Submit" />
       </form>
     </div>
 
   </div>
 
-=======
-<nav class='navbar'>
-    <h1>Offline Beats</h1>
-    <h4>
-        {{!-- <a href='/login'>Login</a> --}}
-    </h4>
->>>>>>> f697fc7a61d72cf4ee72c6f0a4fae85eca36cff4
 </nav>
 
 <div class='container intro'>
-    <h5 class='col-11 col-sm-10 col-md-10 col-lg-9 col-xl-9'>
-        This spotify library app is an online tool to help your music stay offline. We will help you back up your music
-        to other formats that are outside of spotify and can be saved directly to your computer. It's the perfect tool
-        to keep your music safe in the event that spotify is unavailable. Who knows what could happen? Along the way, we
-        will show you stats about your playlists' music, such as the danceability and the energy of the playlist. Click
-        the button below to login to spotify and see your playlists!
-    </h5>
-    <!-- Need to be logged in first before you see this button -->
-    <a href="/library"><button id="seePlaylists-btn" class='seePlaylists'>See My Playlists!</button></a><br>
-    {{!-- <a href="/spotify-auth"><button id="spotifyAuth-btn" class='seePlaylists'>Search</button></a> --}}
+  <h5 class='col-11 col-sm-10 col-md-10 col-lg-9 col-xl-9'>
+    This spotify library app is an online tool to help your music stay offline. We will help you back up your music
+    to other formats that are outside of spotify and can be saved directly to your computer. It's the perfect tool
+    to keep your music safe in the event that spotify is unavailable. Who knows what could happen? Along the way, we
+    will show you stats about your playlists' music, such as the danceability and the energy of the playlist. Click
+    the button below to login to spotify and see your playlists!
+  </h5>
+  <!-- Need to be logged in first before you see this button -->
+  <a href="/library"><button id="seePlaylists-btn" class='seePlaylists'>See My Playlists!</button></a><br>
+  {{!-- <a href="/spotify-auth"><button id="spotifyAuth-btn" class='seePlaylists'>Search</button></a> --}}
 </div>
 
 
 
 <script type="text/javascript">
-    $("#seePlaylists-btn").on("click", function () {
-        $.get("/playlist-library");
-    });
+  $("#seePlaylists-btn").on("click", function () {
+    $.get("/playlist-library");
+  });
 
-    $("#spotifyAuth-btn").on("click", function () {
-        window.location.href = "/spotify-auth";
-    });
+  $("#spotifyAuth-btn").on("click", function () {
+    window.location.href = "/spotify-auth";
+  });
 
 </script>
